Create README.md
myMap (tenative title) is an app where you can keep track of all your travels in one place as well as see your friends.

Each user has their own map page, similar to having their own profile page on your standard social media. 

On that page, the user can locate a place they have traveled through searching or manually finding it on the map and clicking, and make a post about it!

They can add a title, description, and a picture for each location on the map they make a post about. 

On that same page is also a feed that shows all the recent posts on the map, as well as comments from their friends!

On the home page, you can browse different users as well as see recent travels posted by all users.